requirements.txt
Pillow==3.4.2
requests==2.12.1
tqdm==4.10.0
